requirements.txt
cffi==1.15.0
cryptography==37.0.2
pycparser==2.21
PyJWT==2.4.0
